pyproject.toml
[project]
name = "forecasting-capacity-planning"
version = "0.1.0"
description = "Demand forecasting translated into capacity planning decisions."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.26",
  "pyyaml>=6.0",
  "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
